% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the cyclestatsv2 Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{options}{Named list of options passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{enableBookmarking}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{uiPattern}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{...}{Additional arguments passed as golem options.}
}
\description{
Main entry point. Initialises data then launches the app.
}
